+++ frontend/config/main.php
@@ -42,7 +42,7 @@ return [
         'view' => [
             'theme' => [
                 'basePath' => '@app/themes/adminltecustom',
-                'baseUrl' => '@web/themes/adminltecustom',
+                'baseUrl' => '@web/themes/adminLtecustom',
                 'pathMap' => [
                     '@app/views' => '@app/themes/adminltecustom',
                 ],