+++ frontend/themes/adminltecustom/site/about.php
@@ -10,7 +10,7 @@ $this->params['breadcrumbs'][] = $this->title;
 <div class="site-about">
     <h1><?= Html::encode($this->title) ?></h1>
 
-    <p>This is Basic Theme page. This is the About page. You may modify the following file to customize its content:</p>
+    <p>This is AdminLTE Theme page. This is the About page. You may modify the following file to customize its content:</p>
 
     <code><?= __FILE__ ?></code>
 </div>