+++ frontend/themes/basic/site/resetPassword.php
@@ -13,7 +13,7 @@ $this->params['breadcrumbs'][] = $this->title;
 <div class="site-reset-password">
     <h1><?= Html::encode($this->title) ?></h1>
 
-    <p>Basic Theme:</p>
+    <p>This is Basic Theme page.</p>
 
     <p>Please choose your new password:</p>